Initial XMR work
Should implement the entire Verifier EXCEPT it doesn't do any work to 
create transactions. I'd like to outsource that to outsource to 
monero-wallet. We could wrap the RingCT code via a C FFI and then use a 
series of RPC routes, yet it'd take way too long for this PoC.

That said, monero-wallet doesn't allow specification of arbitrary 
spend/view keys over the RPC. You have to use the CLI to create the 
wallet. We could manually craft the wallet file, and then use the RPC to 
load it, yet the internal wallet file spec isn't exactly documented.

For now, this just prints the keys when it recovers them.

There's no meaningful work done on client and the verifier has plenty of 
work which likely can be merged to engine. I'm waiting to see what 
functionality the client also uses.