get_tramsactions_in_block and bug fixes.
Currently faults when the verifier is trying to confirm the Monero send 
was published. That could mean anything from the key formulas to the 
commitment generation to... that entire function is a bit monolithic and 
should probably be broken up, with the pieces obtaining their own test 
suite, before moving forward.