Add shielded ZEC support.
On a branch for a few reasons:
1) Code quality.
2) Hasn't been reviewed by anyone experienced with ZCash.
3) Adds 50 MB to the product.

While 2/3 should be explanatory, the first is a lengthier. Beyond 
generic lint, the JubJub engine has a horrible `bytes` solution to 
implement serialization which wasn't applied to the JubJub types. While 
it should PrivateKey([u8; 32]) at worst, that didn't work for some 
reason. It may have been a cascade of other problems being worked 
through at the time though.

The diversifier is generated as a product of nsk with a DST, instead of 
being generated as it should be. This was a quick 'optimization', and 
potentially de-anonymizing now that I think about it. It depends on if 
ZIP32 and the SNARKS solution provide a clear relation between ak/nk and 
the diversifier. It shouldn't, but if it does, this diversifier won't 
hold to the same relation. No matter what, it should be replaced with a 
proper algorithm (one described in the protocol documentation OR 
something mimicing ZIP32 as best as possible).

The diversifier is also generated in get_deposit_address. It should be 
generated in set_ak_and_ask, removing a stub call from Verifier's claim.

It does work though. It doesn't have a CI hooked up, yet neither does 
Meros/Monero, so that wouldn't be considered a blocker for this PoC of a 
repo.